requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==4.4.2
dotenv==0.9.9
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
instaloader==4.14.1
moviepy==1.0.3
multidict==6.4.4
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pip-autoremove==0.10.0
pluggy==1.6.0
proglog==0.1.12
propcache==0.3.1
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.3
setuptools==80.9.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
